Migrating the module to the Terraform Plugin SDK.
This migrates the module to the Terraform Plugin SDK using the automated
migration tool.

Been tested locally against the terraform provider that uses this and
the provider passes tests with this included.

It would be nice if after this merge the repo could be tagged v0.0.1 so
the version being called from the Terraform provider to be a straight
version. Hunting down this branch and commit that was currently being
used took a few seconds.