Returning error when HTTP response is not OK
If HTTP response is not OK, "Do" method should return an error.
Otherwise, the caller does not get notified that something went wrong.
For instance, if user credentials is wrong for connecting to an APIC, the following was
returned when APIC was queried for some MOs:
    strconv.ParseInt: parsing "{}": invalid syntax
This is because the caller was expecting a certain response body format, and it didn't know
something went wrong. And the error message is misleading to the end-user.

After applying the changes, the following is what end-user sees:
    401 Unauthorized. Response body: {"totalCount":"1","imdata":[{"error":{"attributes":{"code":"401","text":"User credential is incorrect - FAILED local authentication"}}}]}